+++ R/cross_correlation.R
@@ -4,7 +4,8 @@
 #' @usage cross_correlation(X, wl = 512, bp = "pairwise.freq.range", ovlp = 70, 
 #' dens = NULL, wn = 'hanning', cor.method = "pearson", parallel = 1, 
 #' path = NULL, pb = TRUE, na.rm = FALSE, cor.mat = NULL, output = "cor.mat", 
-#' compare.matrix = NULL, type = "spectrogram", nbands = 40, method = 1)
+#' templates = NULL, surveys = NULL, compare.matrix = NULL, type = "fourier",
+#'  nbands = 40, method = 1)
 #' @param  X 'selection_table', 'extended_selection_table' or data frame containing columns for sound files (sound.files), 
 #' selection number (selec), and start and end time of signal (start and end). 
 #' All selections must have the same sampling rate. 
@@ -27,11 +28,13 @@
 #' @param na.rm Logical. If \code{TRUE} all NAs produced when pairwise cross-correlations failed are removed from the 
 #' results. This means that all selections with at least 1 cross-correlation that failed are excluded.
 #' @param cor.mat DEPRECATED. Use 'compare.matrix' instead.
-#' @param output Character vector of length 1 to determine if only the correlation matrix is returned ('cormat') or a list ('list') containing 1) the correlation matrix and 2) a data frame with correlation values at each sliding step for each comparison. The list, which is also of class 'xcorr.output', can be used to find detection peaks with \code{\link{find_peaks}} or to graphically explore detections using \code{\link{lspec}}.
-#' @param compare.matrix A character matrix with 2 columns indicating the selections to be compared (column 1 vs column 2). The columns must contained the ID of the selection, which is given by combining the 'sound.files' and 'selec' columns of 'X',  separated by '-' (i.e. \code{paste(X$sound.files, X$selec, sep = "-")}). Default is \code{NULL}. If supplied only those comparisons will be calculated (as opposed to all pairwise comparisons as the default behavior) and the output will be a data frame composed of the supplied matrix and the correspondent cross-correlation values. Note that 'method' is automatically set to 2 (create spectrograms on the fly) when 'compare.matrix' is supplied but can be set back to 1.
-#' @param type A character vector of length 1 specifying the type of cross-correlation; either "spectrogram" (i.e. spectrographic cross-correlation using Fourier transform; internally using \code{\link[seewave]{spectro}}; default) or "mfcc" (Mel cepstral coefficient cross-correlation; internally using \code{\link[tuneR]{melfcc}}).
+#' @param output Character vector of length 1 to determine if only the correlation matrix is returned ('cor.mat', default) or a list ('list') containing 1) the correlation matrix and 2) a data frame with correlation values at each sliding step for each comparison. The list, which is also of class 'xcorr.output', can be used to find detection peaks with \code{\link{find_peaks}} or to graphically explore detections using \code{\link{lspec}}.
+#' @param templates Character vector with the selections in 'X' to be used as templates for cross-correlation detection. To refer to specific selections in 'X' the user must use the format "sound.file-selec" (e.g. "file1.wav-1"). If only the sound file name is included then the entire sound file is used as template.
+#' @param surveys Character vector with the selections in 'X' to be used as surveys for cross-correlation detection. To refer to specific selections in 'X' the user must use the format "sound.file-selec" (e.g. "file1.wav-1"). If only the sound file name is included then the entire sound file is used as survey.
+#' @param compare.matrix A character matrix with 2 columns indicating the selections to be compared (column 1 vs column 2). The columns must contained the ID of the selection, which is given by combining the 'sound.files' and 'selec' columns of 'X',  separated by '-' (i.e. \code{paste(X$sound.files, X$selec, sep = "-")}). Default is \code{NULL}. If the 'selec' ID is not supplied then the entire sound file is used instead. If 'compare.matrix' is supplied only those comparisons will be calculated (as opposed to all pairwise comparisons as the default behavior) and the output will be a data frame composed of the supplied matrix and the correspondent cross-correlation values. Note that 'method' is automatically set to 2 (create spectrograms on the fly) when 'compare.matrix' is supplied but can be set back to 1. When supplied 'output' is forced to be 'list'. 
+#' @param type A character vector of length 1 specifying the type of cross-correlation; either "fourier" (i.e. spectrographic cross-correlation using Fourier transform; internally using \code{\link[seewave]{spectro}}; default) or "mel" (Mel spectrogram cross-correlation; internally using \code{\link[tuneR]{melfcc}}).
 #' @param nbands Numeric vector of length 1 controlling the number of warped spectral bands to calculate when using \code{type = "mfcc"} (see \code{\link[tuneR]{melfcc}}). Default is 40. 
-#' @param method Numeric vector of length 1 to control the method used to create spectrogram (or mfcc) matrices. Two option are available:
+#' @param method Numeric vector of length 1 to control the method used to create spectrogram matrices. Two option are available:
 #' \itemize{
 #'    \item \code{1}:  matrices are created first (keeping them internally as a list) and cross-correlation is calculated on a second step. Note that this method may require lots of memory if selection and or sound files are large. 
 #'    \item \code{2}: matrices are created "on the fly" (i.e. at the same time that cross-correlation is calculated). More memory efficient but may require extracting the same matrix several times, which will affect performance. Note that when using this method the function does not check if sound files have the same sampling rate which if not, may produce an error.
@@ -41,8 +44,8 @@
 #' the maximum correlation for each pairwise comparison ('max.xcorr.matrix') and 2) a data frame with the correlation statistic for each "sliding" step ('scores').
 #' @export
 #' @name cross_correlation
-#' @details This function calculates the pairwise similarity of multiple signals by means of time-frequency cross-correlation. Spectrographic cross-correlation (SPCC, i.e. Fourier transform) and Mel frequency cepstral coefficients (mfcc) can be applied to create time-frequency representations of sound. 
-#' This method "slides" the spectrogram of the sorthest selection over the longest one calculating a correlation of the amplitude values at each step.
+#' @details This function calculates the pairwise similarity of multiple signals by means of time-frequency cross-correlation. Fourier or Mel spectrograms can be used as time-frequency representations of sound. 
+#' This method "slides" the spectrogram of the shortest selection over the longest one calculating a correlation of the amplitude values at each step.
 #' The function runs pairwise cross-correlations on several signals and returns a list including the correlation statistic
 #' for each "sliding" step as well as the maximum (peak) correlation for each pairwise comparison. To accomplish this the margins
 #' of the signals are expanded by half the duration of the signal both before and after the provided time coordinates. 
@@ -89,7 +92,8 @@
 cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ovlp = 70, dens = NULL, 
                   wn ='hanning', cor.method = "pearson", parallel = 1, 
                   path = NULL, pb = TRUE, na.rm = FALSE, cor.mat = NULL, output = "cor.mat",
-                  compare.matrix = NULL, type = "spectrogram", nbands = 40, method = 1)
+                  templates = NULL, surveys = NULL,
+                  compare.matrix = NULL, type = "fourier", nbands = 40, method = 1)
 {
   # set pb options 
   on.exit(pbapply::pboptions(type = .Options$pboptions$type), add = TRUE)
@@ -159,7 +163,8 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
     if (!is.vector(ovlp)) stop("'ovlp' must be a numeric vector of length 1") else{
       if (!length(ovlp) == 1) stop("'ovlp' must be a numeric vector of length 1")}} 
   
-  if (!is_extended_selection_table(X)){
+  # extended selection table only need sound files in the working directory when doing detection
+  if (!is_extended_selection_table(X) | is_extended_selection_table(X) & !is.null(templates) | is_extended_selection_table(X) & !is.null(compare.matrix)){
     #return warning if not all sound files were found
     fs <- list.files(path = path, pattern = "\\.wav$|\\.wac$|\\.mp3$|\\.flac$", ignore.case = TRUE)
     if (length(unique(X$sound.files[(X$sound.files %in% fs)])) != length(unique(X$sound.files))) 
@@ -185,9 +190,28 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
   # add selection id column to X
   X$selection.id <- paste(X$sound.files, X$selec, sep = "-")
   
-  # keep only selections in supplied compare.matrix to improve performance
+  # if templates were supplied
+  if (!is.null(templates))
+    if (!all(templates %in% X$selection.id))
+      stop("not all templates are referenced in 'X'")
+  
+  # check surveys
+  if (!all(surveys %in% fs))
+    stop("not all sound files in 'surveys' are found in the working directory or 'path'")
+  
+  # if templates were supplied but no compare.matrix use files in working directory
+  if (!is.null(templates) & is.null(compare.matrix))
+    if (is.null(surveys))
+      compare.matrix <- data.frame(X1 = rep(templates, length(fs)), X2 = fs) else
+        compare.matrix <- data.frame(X1 = rep(templates, length(surveys)), X2 = surveys)
+  
+  
   if (!is.null(compare.matrix))
-  { X <- X[X$selection.id %in% unique(c(compare.matrix)), , drop = FALSE]
+    output <- "list"
+      
+  # keep only selections in supplied compare.matrix to improve performance
+  if (!is.null(compare.matrix)){ 
+    X <- X[X$selection.id %in% unique(unlist(c(compare.matrix))), , drop = FALSE]
  
   # set method to 2 if not provided in call
   if (!any(names(call.argms) == "method"))
@@ -202,24 +226,26 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
         max.stps <- 1
   } 
   
-  # generate all possible combinations of selections, keep one with the orignal order of rows to create cor.table output
+  # generate all possible combinations of selections, keep one with the original order of rows to create cor.table output
   if (is.null(compare.matrix))
     spc.cmbs.org <- spc.cmbs <- t(combn(X$selection.id, 2)) else
     {
-      # if all are selection in compare.matrix are from X
-      if (all(c(compare.matrix) %in% X$selection.id))
+      # if all selections in compare.matrix are from X
+      if (all(unlist(c(compare.matrix)) %in% X$selection.id))
         spc.cmbs.org <- spc.cmbs <- compare.matrix else {
           
           # if some are complete sound files
-          complt.sf <- setdiff(c(compare.matrix), X$selection.id)
+          entire.sf <- setdiff(unlist(c(compare.matrix)), X$selection.id)
           
           # get duration of files
-          wvdr <- duration_wavs(files = complt.sf, path = path)
+          wvdr <- info_wavs(path = path)
+          
+          wvdr <- wvdr[wvdr$sound.files %in% entire.sf, ]
           
           # put it in a data frame
           names(wvdr)[2] <- "end"
           wvdr$start <- 0
-          wvdr$selec <- "whole.file"    
+          wvdr$selec <- "entire.file"    
           wvdr$selection.id <- paste(wvdr$sound.files, wvdr$selec, sep = "-")
           
           out <- lapply(1:nrow(wvdr), function(x) {
@@ -242,30 +268,37 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
           
           wvdr <- do.call(rbind, out)        
           
+          if (length(unique(wvdr$sample.rate)) > 1)
+            stop("not all sound files have the same sampling rate")
+          
+          if (is_extended_selection_table(X))  
+            if (unique(wvdr$sample.rate) != unique(attr(X, "check.results")$sample.rate))
+              stop("not all sound files have the same sampling rate than the wave objects in the extended selection table")
+          
           #get intersect of column names
           int.nms <- intersect(names(X), names(wvdr))
             
           X <- rbind(as.data.frame(X)[, int.nms, drop = FALSE], wvdr[, int.nms])
           
           # change complete file names in compare matrix
-          for (i in complt.sf)
-            compare.matrix[compare.matrix == i] <- paste(i, "whole.file", sep = "-")
+          for (i in entire.sf)
+            compare.matrix[compare.matrix == i] <- paste(i, "entire.file", sep = "-")
           
           # set new matrices to allow changes down stream
           spc.cmbs.org <- spc.cmbs <- compare.matrix
         }
       
     }
-  
+
   # function to get spectrogram or mfcc matrices
   spc_FUN <- function(j, pth, W, wlg, ovl, w, nbnds) {
     
     clp <- warbleR::read_sound_file(X = W, index = j, path = pth)
     
-    if (type == "spectrogram")
+    if (type == "fourier" | type == "spectrogram")
       spc <- seewave::spectro(wave = clp, wl = wlg, ovlp = ovl, wn = w, plot = FALSE, fftw = TRUE, norm = TRUE)
     
-    if (type == "mfcc")
+    if (type == "mfcc"| type == "mel")
     {
       # calculate MFCCs
       spc <- melfcc(clp, nbands = nbnds, hoptime =  (wlg / clp@samp.rate) * (ovl / 100), wintime =  wlg / clp@samp.rate, dcttype = "t3", fbtype = "htkmel", spec_out = TRUE)
@@ -289,9 +322,7 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
   if (method == 1){
   #create spectrograms
   if (pb) 
-    if (type == "spectrogram")
-      write(file = "", x = paste0("creating spectrogram matrices (step 1 of ", max.stps,"):")) else
-        write(file = "", x = paste0("creating MFCC matrices (step 1 of ", max.stps,"):"))
+      write(file = "", x = paste0("creating spectrogram matrices (step 1 of ", max.stps,"):"))
   
   
   # set pb options 
@@ -411,7 +442,7 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
   } 
 } else {
   mat <- data.frame(compare.matrix, score = mx.xcrrs)
- if (na.rm) mat <- mat[!is.infinite(mat$scores), ]
+ if (na.rm) mat <- mat[!is.infinite(mat$score), ]
   }
   ## create correlation data matrix (keeps all correlation values, not only the maximum)
   # time is the one of the longest selection as the shortest is used as template 
@@ -425,11 +456,12 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
         )
       
       # put in a data frame
-      df <- data.frame(dyad = paste(spc.cmbs.org[x, ], collapse = "/"),
+      df <- suppressWarnings(data.frame(dyad = paste(spc.cmbs.org[x, ], collapse = "/"),
                        sound.files = spc.cmbs.org[x, which.max(durs)], 
                        template = spc.cmbs.org[x, which.min(durs)], 
                        time = if (!is.null( xcrrs[[x]])) c(X$start[X$selection.id == spc.cmbs.org[x, 1]], X$start[X$selection.id == spc.cmbs.org[x, 2]])[which.max(durs)] + seq(min(durs)/2, max(durs) - min(durs)/2, length.out = length(xcrrs[[x]])) else NA, 
-                       score = if (!is.null(xcrrs[[x]])) xcrrs[[x]] else NA)
+                       score = if (!is.null(xcrrs[[x]])) xcrrs[[x]] else NA
+                      ))
           
       return(df)
       })
@@ -461,6 +493,7 @@ cross_correlation <- function(X = NULL, wl = 512, bp = "pairwise.freq.range", ov
       errors = if (na.rm) errors else NA,
       parameters = lapply(call.argms, eval),
       call = base::match.call(),
+      spectrogram = type,
       warbleR.version = packageVersion("warbleR")
       )
     
@@ -503,11 +536,11 @@ print.xcorr.output <- function(x, ...) {
   cat(crayon::silver(crayon::italic(gsub("    ", "", cll), "\n")))
   
   max.scores <- x$max.xcorr.matrix
-  max.scores$X2 <- gsub("-whole.file", "", max.scores$X2)
+  max.scores$X2 <- gsub("-entire.file", "", max.scores$X2)
   max.scores$score <- NULL
   names(max.scores) <- c("template", "survey")
   
-  cat(crayon::silver(paste("\n The routine was run using the following templates and surveys (files in which templates were searched for): \n")))
+  cat(crayon::silver(paste("\n The routine was run on", x$spectrogram, "spectrograms using the following templates and surveys (selections or files in which templates were looked for): \n")))
 
   #print data frame
   #define columns to show
@@ -522,14 +555,16 @@ print.xcorr.output <- function(x, ...) {
 
   cat(crayon::silver(paste(crayon::bold("\nIncludes:"), "\n* A data frame ('max.xcorr.matrix') with the highest correlation value for each pair of templates and surveys \n")))
   
-  cat(crayon::silver("\n* A data frame ('scores',", nrow(max.scores), "rows) with the cross correlation scores for each of the", nrow(max.scores), "template/survey combination(s) \n"))
+  cat(crayon::silver("\n* A data frame ('scores',", nrow(x$scores), "rows) with the cross correlation scores for each of the", nrow(max.scores), "template/survey combination(s) \n"))
   cat(crayon::silver(paste("\n* A selection table data frame ('org.selection.table') referencing the templates location in sound files \n")))
+  
+  cat(crayon::silver(paste("\n Use", crayon::bold(crayon::italic("full_spectrograms()")), "to plot cross_correlation scores along spectrograms \n")))
+  
+  cat(crayon::silver(paste("\n Use", crayon::bold(crayon::italic("find_peaks()")), "to extract detections from this object \n")))
   
   # print warbleR version
   if (!is.null(x$warbleR.version))
     cat(crayon::silver(paste0("\n Created by warbleR ", x$warbleR.version)), "\n") else
       cat(crayon::silver("\n Created by warbleR < 1.1.27 \n"))
   
-  cat(crayon::silver(paste("\n Use", crayon::bold(crayon::italic("find_peaks()")), "to extract detections from this object \n")))
-
-  }
+}